refactor(db): remove unexported "dB" type
The dB type isn't necessary. All its methods have been refactored to the
exported functions in "db.go".
